IGNITE-14250 .NET: Fix race condition in Events example
* Wait for events to arrive before unsubscribing
* Move local listener from shared project - it does not need to be shared
